+++ src/pages/popup/index.js
@@ -9,7 +9,7 @@ import sendRuntimeMessage$ from "../../core/events/internal/runtime-send-message
 import {EXTENSION_MESSAGES} from "../../config/background";
 
 
-const {POPUP_STATUS, AUTO_SAVE_STATUS, PRICE_UPDATE_STATUS} = EXTENSION_MESSAGES;
+const {POPUP_STATUS, AUTO_SAVE_STATUS, PRICE_UPDATE_STATUS, STOP_FOLLOW_STATUS} = EXTENSION_MESSAGES;
 const BUTTON_STATUS = {
     active: "active",
     inactive: "inactive",
@@ -52,6 +52,23 @@ function updatePriceUpdateButton(buttonEnabled) {
     }
 }
 
+function updateStopFollowButton(buttonEnabled) {
+    if (buttonEnabled === true) {
+        this.setState({
+            stopFollowButtonStatus: BUTTON_STATUS.active
+        });
+    } else if (buttonEnabled === false) {
+        this.setState({
+            stopFollowButtonStatus: BUTTON_STATUS.inactive
+        });
+    }
+}
+
+function onItemChangedTrackStatus({isAutoSaveEnabled, isStopFollowEnabled}) {
+    this.updateAutosaveButton(isAutoSaveEnabled);
+    this.updateStopFollowButton(isStopFollowEnabled);
+}
+
 async function setPendingAutosave() {
     await this.setState({
         autosaveButtonStatus: BUTTON_STATUS.pending
@@ -61,6 +78,12 @@ async function setPendingAutosave() {
 async function setPendingPriceUpdate() {
     await this.setState({
         priceUpdateButtonStatus: BUTTON_STATUS.pending
+    });
+}
+
+async function setPendingStopFollow() {
+    await this.setState({
+        stopFollowButtonStatus: BUTTON_STATUS.pending
     });
 }
 
@@ -76,21 +99,31 @@ class Popup extends Component {
         this.updateAutosaveButton = updateAutosaveButton.bind(this);
         this.setPendingAutosave = setPendingAutosave.bind(this);
         this.updatePriceUpdateButton = updatePriceUpdateButton.bind(this);
+        this.updateStopFollowButton = updateStopFollowButton.bind(this);
         this.setPendingPriceUpdate = setPendingPriceUpdate.bind(this);
+        this.setPendingStopFollow = setPendingStopFollow.bind(this);
+        this.onItemChangedTrackStatus = onItemChangedTrackStatus.bind(this);
 
         this.state = {
             recordButtonStatus: BUTTON_STATUS.inactive,
             autosaveButtonStatus: BUTTON_STATUS.inactive,
-            priceUpdateButtonStatus: BUTTON_STATUS.inactive
+            priceUpdateButtonStatus: BUTTON_STATUS.inactive,
+            stopFollowButtonStatus: BUTTON_STATUS.inactive
         };
 
-        this.triggerPopupStatus$().pipe(
+        const onPopupStatus$ = this.triggerPopupStatus$();
+
+        onPopupStatus$.pipe(
             switchMap(statuses => this.triggerAutoSaveStatus$(statuses))
         ).subscribe(this.updateAutosaveButton);
 
-        this.triggerPopupStatus$().pipe(
+        onPopupStatus$.pipe(
             switchMap(statuses => this.triggerPriceUpdateStatus$(statuses))
         ).subscribe(this.updatePriceUpdateButton);
+
+        onPopupStatus$.pipe(
+            switchMap(statuses => this.triggerStopFollowStatus$(statuses))
+        ).subscribe(this.updateStopFollowButton);
     }
 
     render() {
@@ -99,8 +132,9 @@ class Popup extends Component {
                 <Toolbar recordButtonStatus={this.state.recordButtonStatus}
                          autosaveButtonStatus={this.state.autosaveButtonStatus}
                          priceUpdateButtonStatus={this.state.priceUpdateButtonStatus}
+                         stopFollowButtonStatus={this.state.stopFollowButtonStatus}
                          onPopupStatus={this.onPopupStatus}
-                         onAutosaveStatus={this.updateAutosaveButton}
+                         onItemChangedTrackStatus={this.onItemChangedTrackStatus}
                          onPriceUpdateStatus={this.updatePriceUpdateButton}
                 />
                 <IconLink href="tracked-items.html" icon="tracked-items" title="Tracked items"/>
@@ -118,6 +152,7 @@ class Popup extends Component {
         }
 
         this.setPendingPriceUpdate();
+        this.setPendingStopFollow();
     }
 
     triggerPopupStatus$() {
@@ -125,7 +160,8 @@ class Popup extends Component {
             tap(this.onPopupStatus),
             map(({state}) => ({
                 autoSaveEnabled: state.autoSaveEnabled,
-                isPriceUpdateEnabled: state.isPriceUpdateEnabled
+                isPriceUpdateEnabled: state.isPriceUpdateEnabled,
+                isStopFollowEnabled: state.isStopFollowEnabled
             })),
             share()
         );
@@ -147,6 +183,15 @@ class Popup extends Component {
             switchMap(() => queryActiveTab$()),
             switchMap(([{id}]) =>
                 sendRuntimeMessage$({type: PRICE_UPDATE_STATUS, payload: {id}})
+            )
+        );
+    }
+
+    triggerStopFollowStatus$({isStopFollowEnabled}) {
+        return of(isStopFollowEnabled).pipe(
+            switchMap(() => queryActiveTab$()),
+            switchMap(([{id}]) =>
+                sendRuntimeMessage$({type: STOP_FOLLOW_STATUS, payload: {id}})
             )
         );
     }